refactor: ideEditor 컴포넌트 디자인 시스템 적용
